fixing bug with is-admin api and adding jwt_token to api calls in the frontend

+++ frontend/pvtlp-react/src/tables/TransactionsTable.js
@@ -5,7 +5,7 @@ import {BackgroundGradient} from "../components/background-gradient";
 import {Flex, IconButton, Button, Badge, AlertDialog} from "@radix-ui/themes";
 import {Pencil1Icon as EditIcon, PlusIcon, PersonIcon, DimensionsIcon, ReloadIcon} from "@radix-ui/react-icons";
 import { TrashIcon } from '@heroicons/react/24/outline';
-import {getCallToBackend, deleteCallToBackend, getJwtToken, isAdminGetCallToBackend} from "../utils/api_call_backend";
+import {getCallToBackend, deleteCallToBackend, isAdminGetCallToBackend} from "../utils/api_call_backend";
 import Flag from "react-world-flags";
 import {API_Endpoint} from "../utils/api_endpoints";
 import {useNavigate} from "react-router-dom";
@@ -61,8 +61,9 @@ export default function TransactionsTable() {
                     getCallToBackend(API_Endpoint.Titles, Title)
                 ]);
 
-                setIsAdmin(isAdmin);
+                setIsAdmin(isAdmin === "true");
                 console.log("isAdmin: " + isAdmin);
+
 
                 setTransactions(transactions);
                 console.log("transactions: " + transactions);
@@ -78,6 +79,11 @@ export default function TransactionsTable() {
         };
         fetchData().then(() => console.log("Data fetched successfully"));
     }, []);
+
+    // when isAdmin value is changed, log it
+    useEffect(() => {
+        console.log("isAdmin value changed: " + isAdmin);
+    }, [isAdmin]);
 
     const rowsPerPage = 4;
 
@@ -96,7 +102,7 @@ export default function TransactionsTable() {
             <Flex direction={"column"} align={"center"}>
                 <Flex direction={'row'} justify={'center'}>
                     <PersonIcon/>
-                    // this to make sure the name is not broken into two lines, equivalent to &nbsp;
+                     {/*this to make sure the name is not broken into two lines, equivalent to &nbsp;*/}
                     <span>{`${customer.first_name}\u00A0${customer.last_name}`}</span> </Flex>
                 <div className="text-gray-500"> Home Region: {customer.home_country_code_fk}</div>
                 <div className="text-gray-500"> ID: {customer.customer_id_pk}</div>
@@ -222,6 +228,13 @@ export default function TransactionsTable() {
             const handleDeleteClick = () => {
                 console.log(`Delete clicked for transaction ID: ${transactionId}`);
                 // api call to deleteCallToBackend() with transactionId
+                if (!isAdmin) {
+                    console.log("logging is-admin from delete function GG" + isAdmin)
+                    setCalloutMessage("You are not authorized to delete transactions");
+                    setCalloutColor("red");
+                    setIsCalloutVisible(true);
+                    return;
+                }
                 deleteCallToBackend(API_Endpoint.Delete_Transaction, transactionId)
                     .then((response) => {
                         console.log("Transaction deleted call was sent successfully: ", response);
